requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
duckduckgo-search
validators
youtube_transcript_api
unstructured
pytube
numexpr
cassio
pypdf2
gradio
langchain_nvidia_ai_endpoints
crewai
crewai-tools